Dynamically determining repo to publish to based on whether it is a release version or not
Maven version strings must not contain '/', as branch names are being included in the version and branch names tends to have slashes, we need to replace this with underscores

The MavenArtifactRepositories that are returned through the standard repository {} elements are different types to those that need to be specified in the publishing.repositories section so need to be recreated (The publishing.repositories silently fails if you pass it invalid data).

We could remove the `version instanceof uk.co.jamesridgway.gradle.gitflow.plugin.version.ReleaseVersion` check if we added a flag in the git flow plugin.